♻️ Migrate the ClickUp operations into a single module to converge relative logics.
